requirements.txt
asgiref==3.6.0
attrs==23.1.0
certifi==2023.5.7
Django==4.2
django-debug-toolbar==4.1.0
django-filter==23.2
djangorestframework==3.14.0
drf-spectacular==0.26.3
inflection==0.5.1
jsonschema==4.17.3
Pillow==9.0.1
platformdirs==3.2.0
pyrsistent==0.19.3
pytz==2023.3
PyYAML==6.0
sentry-sdk==1.27.0
sqlparse==0.4.4
uritemplate==4.1.1
urllib3==2.0.3
